Updated path and renamed artefact from 'testando' to 'gerenciador'
