+++ src/main/java/com/gerenciador/servlet/AlteraEmpresaServlet.java
@@ -1,4 +1,4 @@
-package com.example.testando;
+package com.gerenciador.servlet;
 
 import javax.servlet.*;
 import javax.servlet.http.*;
