+++ src/main/java/com/gerenciador/servlet/Empresa.java
@@ -1,4 +1,4 @@
-package com.example.testando;
+package com.gerenciador.servlet;
 
 import java.util.Date;
