+++ src/main/java/com/gerenciador/servlet/NovaEmpresaServlet.java
@@ -1,10 +1,9 @@
-package com.example.testando;
+package com.gerenciador.servlet;
 
 import javax.servlet.*;
 import javax.servlet.http.*;
 import javax.servlet.annotation.*;
 import java.io.IOException;
-import java.io.PrintWriter;
 import java.text.ParseException;
 import java.text.SimpleDateFormat;
 import java.util.Date;
