+++ src/main/java/com/gerenciador/servlet/OiServlet.java
@@ -1,4 +1,4 @@
-package com.example.testando;
+package com.gerenciador.servlet;
 
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
